fix importo

+++ src/main/java/carmine/pittella/home/model/entity/MovimentoEntity.java
@@ -5,6 +5,7 @@ import io.quarkus.hibernate.orm.panache.PanacheEntityBase;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -46,7 +47,7 @@ public class MovimentoEntity extends PanacheEntityBase {
     public String titolo;
 
     @Column(name = "importo", nullable = false, precision = 10, scale = 2)
-    public Long importo;
+    public BigDecimal importo;
 
     @Column(name = "commento")
     public String commento;
